# YJIT: Optimize rb_int_equal